Pulled all simType trimming/pre-rendering out into its own module
