+++ src/sass/styles.scss.d.ts
@@ -21,7 +21,7 @@ export const content: string;
 export const contentDarkPlus: string;
 export const simTypeDarkPlus: string;
 export const simTypeCaret: string;
-export const blinking: string;
+export const blinkingDarkPlus: string;
 export const simTypeQuoting: string;
 export const typing: string;
 export const lineNumber: string;
@@ -41,3 +41,4 @@ export const tintDefault: string;
 export const simTypeCapoLight: string;
 export const tabsCapoLight: string;
 export const contentCapoLight: string;
+export const blinkingCapoLight: string;
